pyproject.toml
[project]
name = "rayna-backend"
version = "2.0.0"
description = "Rayna Tours AI Chatbot — Python/FastAPI migration"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    # LLM SDKs
    "anthropic==0.42.0",
    "openai==1.58.1",
    "groq==0.13.1",
    # LangGraph agent framework
    "langgraph==0.2.60",
    "langchain-core==0.3.28",
    # MongoDB (async)
    "motor==3.6.0",
    "pymongo==4.9.2",
    # Vector DB
    "pinecone-client==5.0.1",
    # Validation & settings
    "pydantic==2.10.3",
    "pydantic-settings==2.7.0",
    # HTTP client
    "httpx==0.28.1",
    # Rate limiting
    "slowapi==0.1.9",
    # CORS
    "starlette==0.41.3",
    # CSV parsing
    "aiofiles==24.1.0",
    # Background tasks
    "celery[redis]==5.4.0",
    "redis==5.2.1",
    # Retry logic
    "tenacity==9.0.0",
    # Logging
    "structlog==24.4.0",
    # SSE streaming
    "sse-starlette==2.2.1",
    # UUID
    "uuid6==2024.7.10",
]

[project.optional-dependencies]
dev = [
    "ruff==0.8.4",
    "mypy==1.13.0",
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "httpx==0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
